+++ projects/01_fyyur/starter_code/fyyur/static/img/front-splash.jpg
