+++ projects/01_fyyur/starter_code/fyyur/static/js/script.js
