+++ projects/01_fyyur/starter_code/fyyur/templates/pages/home.css
