+++ projects/01_fyyur/starter_code/fyyur/utils.py
@@ -1,0 +1,5 @@
+import os
+
+
+def env(key, default=None):
+    return os.getenv(key, default)
